Certain tweaks to work better in library setting and with missing locations
